# handoff/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Iron Mind API",
        "version": "local"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Local Docker"
        }
    ],
    "paths": {
...
        "/checkout/session": {
            "post": {
                "tags": [
                    "Checkout"
                ],
                "summary": "Create Checkout Session",
                "description": "Create a Stripe Checkout Session.",
                "operationId": "create_checkout_session_checkout_session_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CheckoutRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CheckoutResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    },
                    {
                        "DebugHeaders": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CheckoutRequest": {
                "properties": {
                    "type": {
                        "type": "string",
                        "enum": [
                            "one_time",
                            "subscription"
                        ],
                        "title": "Type"
                    },
                    "courseId": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Courseid"
                    }
                },
                "type": "object",
                "required": [
                    "type"
                ],
                "title": "CheckoutRequest"
            },
            "CheckoutResponse": {
                "properties": {
                    "url": {
                        "type": "string",
                        "title": "Url"
                    }
                },
                "type": "object",
                "required": [
                    "url"
                ],
                "title": "CheckoutResponse"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            },
            "DebugHeaders": {
                "type": "apiKey",
                "in": "header",
                "name": "X-Debug-Uid",
                "description": "DEV ONLY: Debug auth UID header"
            },
            "DebugAdmin": {
                "type": "apiKey",
                "in": "header",
                "name": "X-Debug-Admin",
                "description": "DEV ONLY: set to '1' to impersonate admin"
            }
        }
    }
}